Fix strange crash with ColourComponent static callback TurboState has width/height

+++ ecs/Entity.hx
@@ -167,7 +167,10 @@ class Entity
             var clr = c.colour;
             
             this.remove(ColourComponent);
-            ColourComponent.onRemove(c);
+            if (ColourComponent.onRemove != null)
+            {
+                ColourComponent.onRemove(c);
+            }
             
             this.add(new ColourComponent(clr.red, clr.green, clr.blue, width, height, this));   
         }
